update now includes speechUpdate
